requirements.txt
apache-beam[gcp]
google-cloud-bigquery
google-cloud-storage
searchtweets-v2